feat: added type and moved type to common folder
